Add files via upload
Here are the updated python, each of those has a read and write file. So the Game will save the message on a local text file, and MLCode will read it, vice-versa. This is the only way I found work so far and one of the concerns I have is the speed, but we can test at the first place. If you guys have any other ideas on how to let two python files passing values to each other please let me know!!!